pyproject.toml
[build-system]
requires = ["setuptools>=80.3.1", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "splent_framework"
version = "1.1.1"
description = "SPLENT-FRAMEWORK is a set of libraries for agile product development within SPLENT."
readme = "README.md"
requires-python = ">=3.13"
authors = [{ name = "DiversoLab", email = "diversolab@us.es" }]
license-files = ["LICENSE"]

dependencies = [
    "python-dotenv==1.0.1",
    "sqlalchemy==2.0.37",
    "flask==3.1.0",
    "flask_sqlalchemy==3.1.1",
    "flask_migrate==4.1.0",
    "flask_session==0.8.0",
    "flask_mail==0.10.0",
    "flask_login==0.6.3",
    "flask_wtf==1.2.2",
    "redis==5.2.1",
    "pymysql==1.1.1",
    "psutil==6.1.1",
    "pytz==2024.2"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-cov>=6.0",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
filterwarnings = ["ignore::pytest.PytestCollectionWarning"]

[project.urls]
Homepage = "https://github.com/diverso-lab/splent_framework"
Issues = "https://github.com/diverso-lab/splent_framework/issues"
